+++ tools/release-assets/main.go
@@ -1,0 +1,28 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"github.com/hypnotox/agentic-workflows/internal/projector"
+)
+
+func main() {
+	if len(os.Args) != 2 {
+		fmt.Fprintln(os.Stderr, "usage: release-assets <awf.sh-path>")
+		os.Exit(2)
+	}
+	path := os.Args[1]
+	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
+		fail(err)
+	}
+	if err := os.WriteFile(path, projector.PublicLauncher(), 0o755); err != nil {
+		fail(err)
+	}
+}
+
+func fail(err error) {
+	fmt.Fprintln(os.Stderr, "release-assets:", err)
+	os.Exit(1)
+}